chore: move sentry config
